requirements.txt
slack-bolt>=1.20.0
slack-sdk>=3.27.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.126.0
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.7.0
groq>=0.9.0
mistralai>=1.0.0
cohere>=5.5.0
requests>=2.31.0
apscheduler>=3.10.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
cryptography>=42.0.0
